feat: introduce Input primitive as base for FormInput (#665)
