+++ packages/reakit-system-bootstrap/.gitignore
@@ -5,6 +5,7 @@
 /Dialog
 /Form
 /Group
+/Input
 /Menu
 /Popover
 /Separator
