+++ packages/reakit-system-bootstrap/src/Form.ts
@@ -1,9 +1,6 @@
 import { css, cx } from "emotion";
 import { unstable_FormHTMLProps, unstable_FormOptions } from "reakit/Form/Form";
-import {
-  unstable_FormInputHTMLProps,
-  unstable_FormInputOptions,
-} from "reakit/Form/FormInput";
+import { unstable_FormInputOptions } from "reakit/Form/FormInput";
 import {
   unstable_FormMessageHTMLProps,
   unstable_FormMessageOptions,
@@ -21,7 +18,6 @@ import { unstable_getIn } from "reakit/Form/utils/getIn";
 import { useBoxProps as usePaletteBoxProps } from "reakit-system-palette/Box";
 import { useContrast } from "reakit-system-palette/utils/contrast";
 import { useFade } from "reakit-system-palette/utils/fade";
-import { usePalette } from "reakit-system-palette/utils/palette";
 import { useLighten } from "reakit-system-palette/utils/lighten";
 import { BootstrapBoxOptions } from "./Box";
 
@@ -59,38 +55,6 @@ export function useFormInputOptions({
     },
     ...options,
   };
-}
-
-export function useFormInputProps(
-  { unstable_system }: BootstrapFormInputOptions,
-  htmlProps: unstable_FormInputHTMLProps = {}
-): unstable_FormInputHTMLProps {
-  const {
-    style: { backgroundColor, borderColor: originalBorderColor },
-  } = usePaletteBoxProps({ unstable_system });
-
-  const foreground = useContrast(backgroundColor) || "black";
-  const color = useLighten(foreground, 0.3);
-  const primary = usePalette("primary");
-  const borderColor = useFade(foreground, 0.75);
-  const focusBorderColor = useLighten(primary, 0.4);
-
-  const formInput = css`
-    display: block;
-    width: 100%;
-    border-radius: 0.2rem;
-    padding: 0.5em 0.75em;
-    font-size: 100%;
-    border: 1px solid ${originalBorderColor || borderColor};
-    color: ${color};
-    margin: 0 !important;
-
-    &:focus {
-      border-color: ${originalBorderColor || focusBorderColor};
-    }
-  `;
-
-  return { ...htmlProps, className: cx(formInput, htmlProps.className) };
 }
 
 export type BootstrapFormMessageOptions = BootstrapBoxOptions &
