+++ packages/reakit/.gitignore
@@ -9,6 +9,7 @@
 /Form
 /Group
 /Id
+/Input
 /Menu
 /Popover
 /Portal
